+++ compiled/jquery-mobile-angular-adapter-1.1.2-SNAPSHOT.js
@@ -1677,23 +1677,34 @@ factory(window.jQuery, window.angular);
 
     function createEventDirective(directive, eventType) {
         mod.directive(directive, function () {
-            return function (scope, element,attrs) {
+            return function (scope, element, attrs) {
                 var eventHandler = attrs[directive];
                 registerEventHandler(scope, element, eventType, eventHandler);
             };
         });
     }
 
-    var eventDirectives = {ngmTaphold:'taphold', ngmSwipe:'swipe', ngmSwiperight:'swiperight',
-        ngmSwipeleft:'swipeleft',
-        ngmPagebeforeshow:'pagebeforeshow',
-        ngmPagebeforehide:'pagebeforehide',
-        ngmPageshow:'pageshow',
-        ngmPagehide:'pagehide',
-        ngmClick:'vclick'
-    };
-    for (var directive in eventDirectives) {
-        createEventDirective(directive, eventDirectives[directive])
+    // See http://jquerymobile.com/demos/1.2.0/docs/api/events.html
+    var jqmEvents = ['tap', 'taphold', 'swipe', 'swiperight', 'swipeleft', 'vmouseover',
+        'vmouseout',
+        'vmousedown',
+        'vmousemove',
+        'vmouseup',
+        'vclick',
+        'vmousecancel',
+        'orientationchange',
+        'scrollstart',
+        'scrollend',
+        'pagebeforeshow',
+        'pagebeforehide',
+        'pageshow',
+        'pagehide'
+    ];
+    var event, directive, i;
+    for (i=0; i<jqmEvents.length; i++) {
+        event = jqmEvents[i];
+        directive = 'ngm' + event.substring(0, 1).toUpperCase() + event.substring(1);
+        createEventDirective(directive, event);
     }
 
 })(angular);